Wrote Coll_Loss test.
I wrote tests for the 1d and 2d versions of this function. Both were
littered with errors, but appear to work now.

I also standardized some of the names (Max_Derivatives ->
Highest_Order_Derivatives, any instance of "_To_" in a function name is
now "_to_").

I still want to write an Lp loss test. Otherwise, however, everything
looks good. Data_Loss and Generate_Points are the only two functions
without tests... that largely because those functions are fairly simple
or are difficult to test.

Once I write an Lp Loss test, I will have finished everything that I
want to prior to the midterm report.

+++ Code/Points.py
@@ -42,8 +42,8 @@ def Generate_Points(
 
     # Make a tensor to hold all the points.
     Points = torch.empty((Num_Points, Num_Dim),
-                         dtype = numpy.int64,
-                         device = Device);
+                          dtype  = torch.float32,
+                          device = Device);
 
     # Populate the coordinates in Points, one coordinate at a time.
     for j in range(Num_Dim):
